Move `mapTraverseableDecoderA` to `Shelley.TxWits`
